feat: implement new methods
- i'm lazy

+++ src/models/score.rs
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2022 Andriel Ferreira <https://github.com/AndrielFR>
 
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default, Clone, PartialEq)]
 pub struct Score {
-    pub average: f64,
-    pub mean: f64,
+    pub average: i64,
+    pub mean: i64,
 }
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq)]
 pub enum Format {
     Point100,
     Point10Decimal,
